Add CPU time to all files in verbose mode This time currently only accounts for parsing time writing the unit. Input aquisition and read time is not included in this time.

+++ src/srcml/timer.hpp
@@ -43,7 +43,7 @@ public:
 	}
 
 	// time in milliseconds
-	inline double cpu_time_elapsed() {
+	inline long cpu_time_elapsed() {
 		return cpu_time.elapsed() * 10000;
 	}
